+++ src/LowerTriangularArrays/lower_triangular_array.jl
@@ -213,6 +213,8 @@ for f in (:zeros, :ones, :rand, :randn)
             $f(LowerTriangularArray{Float64}, spectrum, I...)
         Base.$f(::Type{LowerTriangularMatrix}, spectrum::AbstractSpectrum) =
             $f(LowerTriangularArray{Float64}, spectrum)
+        Base.$f(spectrum::AbstractSpectrum, I::Vararg{Integer, M}) where M =
+            $f(LowerTriangularArray{Float64}, spectrum, I...)
     end
 end
 
@@ -249,6 +251,8 @@ end
 function LowerTriangularMatrix{T}(::UndefInitializer, spectrum::AbstractSpectrum) where T
     return LowerTriangularMatrix(Vector{T}(undef, nonzeros(spectrum)), spectrum)
 end
+
+Base.eltype(L::LowerTriangularArray) = eltype(L.data)
 
 # INDEXING
 """
@@ -648,6 +652,9 @@ Base.similar(L::LowerTriangularArray{T, N, ArrayType, SP}, ::Type{T}) where {T, 
     LowerTriangularArray{T, N, ArrayType, SP}(similar(L.data, T), L.spectrum)
 Base.similar(L::LowerTriangularArray{T}) where T = similar(L, T)
  
+array_type(::Type{<:LowerTriangularArray{T, N, ArrayType}}) where {T, N, ArrayType} = ArrayType
+array_type(L::LowerTriangularArray) = array_type(typeof(L))
+
 Base.prod(L::LowerTriangularArray{NF}) where NF = zero(NF)
 @inline Base.sum(L::LowerTriangularArray; dims=:, kw...) = sum(L.data; dims, kw...)